Backend sprint 2

+++ backend/src/main/java/be/ucll/mobile/rngenius/generator/service/GeneratorService.java
@@ -1,17 +1,24 @@
 package be.ucll.mobile.rngenius.generator.service;
 
 import java.util.List;
-
+import java.util.ArrayList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import be.ucll.mobile.rngenius.generator.model.Generator;
 import be.ucll.mobile.rngenius.generator.repo.GeneratorRepository;
 import be.ucll.mobile.rngenius.option.model.Option;
 import be.ucll.mobile.rngenius.option.repo.OptionRepository;
+import be.ucll.mobile.rngenius.participant.model.Participant;
+import be.ucll.mobile.rngenius.participant.repo.ParticipantRepository;
+import be.ucll.mobile.rngenius.selection.model.Selection;
+import be.ucll.mobile.rngenius.selection.repo.SelectionRepository;
+import be.ucll.mobile.rngenius.user.model.User;
 import be.ucll.mobile.rngenius.user.service.UserService;
 import be.ucll.mobile.rngenius.user.service.UserServiceException;
+import jakarta.transaction.Transactional;
 
 @Service
+@Transactional
 public class GeneratorService {
 
     @Autowired
@@ -19,6 +26,12 @@ public class GeneratorService {
 
     @Autowired
     private OptionRepository optionRepository;
+
+    @Autowired
+    private ParticipantRepository participantRepository;
+
+    @Autowired
+    private SelectionRepository selectionRepository;
 
     @Autowired
     private UserService userService;
@@ -30,11 +43,15 @@ public class GeneratorService {
 
         if (generator == null) {
             throw new GeneratorServiceException("generator", "No generator with this id");
-        } else if (!generator.getUser().id.equals(requesterId)) {
-            throw new GeneratorServiceAuthorizationException("generator", "You are not authorized to retrieve this generator");
-        }
+        } 
 
-        return generator;
+        for (Participant participant : generator.getParticipants()) {
+            if (participant.getUser().id.equals(requesterId)) {
+                return generator;
+            }
+        } 
+
+        throw new GeneratorServiceAuthorizationException("generator", "You are not authorized to view this generator");
     }
 
     public List<Generator> getMyGenerators(Long requesterId) {
@@ -49,10 +66,37 @@ public class GeneratorService {
         generator.setUser(userService.getUserById(requesterId));
 
         generatorRepository.save(generator);
+
+        Participant participant = new Participant();
+        participant.setGenerator(generator);
+        participant.setUser(generator.getUser());
+        participantRepository.save(participant);
+
+    }
+
+    public void updateGenerator(Long id, Generator generator, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException, UserServiceException {
+        if (generator == null) {
+            throw new GeneratorServiceException("generator", "Generator data is required");
+        }
+
+        Generator existingGenerator = getGeneratorById(id, requesterId);
+
+        if (!existingGenerator.getUser().id.equals(requesterId)) {
+            throw new GeneratorServiceAuthorizationException("generator", "You are not authorized to update this generator");
+        }
+
+        existingGenerator.setTitle(generator.getTitle());
+        existingGenerator.setIconNumber(generator.getIconNumber());
+
+        generatorRepository.save(existingGenerator);
     }
 
     public void deleteGeneratorById(Long id, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException {
         Generator generator = getGeneratorById(id, requesterId);
+
+        if (!generator.getUser().id.equals(requesterId)) {
+            throw new GeneratorServiceAuthorizationException("generator", "You are not authorized to delete this generator");
+        }
 
         generatorRepository.delete(generator);
     }
@@ -77,6 +121,13 @@ public class GeneratorService {
 
         option.setGenerator(generator);
         optionRepository.save(option);
+
+        for (Participant participant : generator.getParticipants()) {
+            Selection selection = new Selection();
+            selection.setParticipant(participant);
+            selection.setOption(option);
+            selectionRepository.save(selection);
+        }
     }
 
     public void deleteCategorizedGeneratorOption(Long optionId, String category, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException {
@@ -91,6 +142,7 @@ public class GeneratorService {
         if (option.getCategories().isEmpty()) {
             optionRepository.delete(option);
         } else {
+            System.out.println(option.getCategories());
             optionRepository.save(option);
         }
     }
@@ -99,14 +151,117 @@ public class GeneratorService {
         Generator generator = getGeneratorById(generatorId, requesterId);
 
         List<Option> options = generator.getOptions();
+        List<Option> validOptions = new ArrayList<>();
 
-        if (options.isEmpty()) {
-            throw new GeneratorServiceException("generator", "No options available");
+        for (Option option : options) {
+            boolean isExcluded = false;
+            int timesFavorised = 0;
+
+            List<Selection> selections = selectionRepository.findSelectionsByOptionId(option.id);
+            for (Selection selection : selections) {
+                if (selection.getExcluded()) {
+                    isExcluded = true;
+                    break;
+                }
+                if (selection.getFavorised()) {
+                    timesFavorised++;
+                }
+            }
+
+            if (!isExcluded) {
+                validOptions.add(option);
+                for (int i = 0; i < timesFavorised; i++) {
+                    validOptions.add(option);
+                }
+            }
         }
 
-        int randomIndex = (int) (Math.random() * options.size());
+        if (validOptions.isEmpty()) {
+            throw new GeneratorServiceException("generator", "No valid options available");
+        }
 
-        return options.get(randomIndex);
+        int randomIndex = (int) (Math.random() * validOptions.size());
+
+        return validOptions.get(randomIndex);
+    }
+    
+    public void favoriseOption(Long optionId, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException {
+        Selection selection = getSelectionByParticipantUserIdAndOptionId(requesterId, optionId);
+        if (!selection.getFavorised()) {
+            selection.setFavorised(true);
+            selection.setExcluded(false);
+        } else {
+            selection.setFavorised(false);
+        }
+        selectionRepository.save(selection);
+    }
+
+    public void excludeOption(Long optionId, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException {
+        Selection selection = getSelectionByParticipantUserIdAndOptionId(requesterId, optionId);
+        if (!selection.getExcluded()) {
+            selection.setFavorised(false);
+            selection.setExcluded(true);
+        } else {
+            selection.setExcluded(false);
+        }
+        selectionRepository.save(selection);
+    }
+
+    public void addGeneratorParticipant(Long generatorId, String email, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException, UserServiceException {
+        Generator generator = getGeneratorById(generatorId, requesterId);
+
+        if (!generator.getUser().id.equals(requesterId)) {
+            throw new GeneratorServiceAuthorizationException("generator", "You are not authorized to add participants to this generator");
+        }
+
+        User user = userService.getUserByEmail(email);
+
+        Participant participant = participantRepository.findParticipantByUserIdAndGeneratorId(user.id, generator.id);
+        if (participant != null) {
+            throw new GeneratorServiceException("participant", "Participant already joined this generator");
+        } else {
+            participant = new Participant();
+            participant.setGenerator(generator);
+            participant.setUser(user);
+            participantRepository.save(participant);
+
+            for (Option option : generator.getOptions()) {
+                Selection selection = new Selection();
+                selection.setParticipant(participant);
+                selection.setOption(option);
+                selectionRepository.save(selection);
+            }
+        }
+    }
+
+    public void removeGeneratorParticipant(Long generatorId, Long participantId, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException {
+        Generator generator = getGeneratorById(generatorId, requesterId);
+
+        if (!generator.getUser().id.equals(requesterId)) {
+            throw new GeneratorServiceAuthorizationException("generator", "You are not authorized to delete participants from this generator");
+        } else if (generator.getUser().id.equals(participantId)) {
+            throw new GeneratorServiceException("participant", "You cannot remove yourself from your own generator");
+        }
+
+        Participant participant = participantRepository.findParticipantByUserIdAndGeneratorId(participantId, generatorId);
+
+        if (participant == null) {
+            throw new GeneratorServiceException("participant", "Participant not found in this generator");
+        }
+
+        participantRepository.delete(participant);
+    }
+
+    public void leaveGenerator(Long generatorId, Long requesterId) throws GeneratorServiceException, GeneratorServiceAuthorizationException {
+        Generator generator = getGeneratorById(generatorId, requesterId);
+
+        if (generator.getUser().id.equals(requesterId)) {
+            throw new GeneratorServiceException("generator", "You cannot leave your own generator");
+        }
+
+        Participant participant = participantRepository.findParticipantByUserIdAndGeneratorId(requesterId, generatorId);
+
+        participantRepository.delete(participant);
     }
 
     private Option getOptionById(Long optionId) throws GeneratorServiceException {
@@ -117,5 +272,15 @@ public class GeneratorService {
         }
 
         return option;
+    }
+
+    private Selection getSelectionByParticipantUserIdAndOptionId(Long participantId, Long optionId) throws GeneratorServiceException {
+        Selection selection = selectionRepository.findSelectionByParticipantUserIdAndOptionId(participantId, optionId);
+
+        if (selection == null) {
+            throw new GeneratorServiceException("selection", "No selection with this participant and option");
+        }
+
+        return selection;
     }
 }
